Refactor runonserver APIs for easier test migration
