--- testsuite/integration-arquillian/servers/auth-server/services/testsuite-providers/src/main/java/org/keycloak/testsuite/runonserver/RunOnServerException.java
@@ -1,12 +1,0 @@
-package org.keycloak.testsuite.runonserver;
-
-/**
- * Created by st on 26.01.17.
- */
-public class RunOnServerException extends RuntimeException {
-
-    public RunOnServerException(Throwable throwable) {
-        super(throwable);
-    }
-
-}
