+++ testsuite/integration-arquillian/tests/base/src/main/java/org/keycloak/testsuite/client/resources/TestingResource.java
@@ -173,16 +173,6 @@ public interface TestingResource {
     @Produces(MediaType.APPLICATION_JSON)
     String getSSOCookieValue();
 
-    @POST
-    @Path("/remove-user-session")
-    @Produces(MediaType.APPLICATION_JSON)
-    void removeUserSession(@QueryParam("realm") final String realm, @QueryParam("session") final String sessionId);
-
-    @POST
-    @Path("/remove-user-sessions")
-    @Produces(MediaType.APPLICATION_JSON)
-    void removeUserSessions(@QueryParam("realm") final String realm);
-
     @GET
     @Path("/get-last-session-refresh")
     @Produces(MediaType.APPLICATION_JSON)
